Refactor db_service from torrent checker and update tests
